+++ src/ft_printf/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: ldulling <ldulling@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/22 19:26:03 by ldulling          #+#    #+#             */
-/*   Updated: 2025/01/30 07:27:55 by ldulling         ###   ########.fr       */
+/*   Updated: 2025/01/30 07:36:50 by ldulling         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 int	ft_printf(const char *format, ...)
 {
 	int			errno_cpy;
-	int			i;
+	size_t		i;
 	size_t		printed;
 	t_format	f;
 	va_list		ap;
