+++ src/ft_printf/print_parsed.c
@@ -6,21 +6,22 @@
 /*   By: ldulling <ldulling@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/24 16:12:14 by ldulling          #+#    #+#             */
-/*   Updated: 2025/01/29 09:48:06 by ldulling         ###   ########.fr       */
+/*   Updated: 2025/01/30 07:44:32 by ldulling         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 #include "libft.h"
+#include <stddef.h>
 
-static int	print_flags(const char *format, int *i, t_format *f);
-static int	print_width(const char *format, int *i, t_format *f);
-static int	print_precision(const char *format, int *i, t_format *f);
+static size_t		print_flags(const char *format, size_t *i, t_format *f);
+static unsigned int	print_width(const char *format, size_t *i, t_format *f);
+static unsigned int	print_precision(const char *format, size_t *i, t_format *f);
 
-int	print_parsed(const char *format, int parsed, t_format *f)
+size_t	print_parsed(const char *format, size_t parsed, t_format *f)
 {
-	int	i;
-	int	printed;
+	size_t	i;
+	size_t	printed;
 
 	i = 0;
 	printed = 0;
@@ -35,9 +36,9 @@ int	print_parsed(const char *format, int parsed, t_format *f)
 	return (printed);
 }
 
-static int	print_flags(const char *format, int *i, t_format *f)
+static size_t	print_flags(const char *format, size_t *i, t_format *f)
 {
-	int	printed;
+	size_t	printed;
 
 	printed = 0;
 	if (f->hash)
@@ -55,9 +56,9 @@ static int	print_flags(const char *format, int *i, t_format *f)
 	return (printed);
 }
 
-static int	print_width(const char *format, int *i, t_format *f)
+static unsigned int	print_width(const char *format, size_t *i, t_format *f)
 {
-	int	printed;
+	unsigned int	printed;
 
 	printed = 0;
 	if (f->width)
@@ -70,9 +71,9 @@ static int	print_width(const char *format, int *i, t_format *f)
 	return (printed);
 }
 
-static int	print_precision(const char *format, int *i, t_format *f)
+static unsigned int	print_precision(const char *format, size_t *i, t_format *f)
 {
-	int	printed;
+	unsigned int	printed;
 
 	printed = 0;
 	if (f->precision >= 0)
