+++ src/ft_printf/print_str.c
@@ -6,41 +6,42 @@
 /*   By: ldulling <ldulling@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/22 19:27:17 by ldulling          #+#    #+#             */
-/*   Updated: 2025/01/29 09:48:16 by ldulling         ###   ########.fr       */
+/*   Updated: 2025/01/30 07:05:32 by ldulling         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 #include "libft.h"
+#include <stddef.h>
 
-static int	print(const char *str, int len, t_format *f);
+static size_t	print(const char *str, size_t len, t_format *f);
 
-int	print_str(const char *str, t_format *f)
+size_t	print_str(const char *str, t_format *f)
 {
-	int	len;
-	int	printed;
+	size_t	len;
+	size_t	printed;
 
 	printed = 0;
 	if (!str)
 	{
 		len = ft_strlen(NULL_PRINTOUT_STR);
-		if (f->precision >= 0 && len > f->precision)
+		if (f->precision >= 0 && len > (size_t)f->precision)
 			len = 0;
 		printed += print(NULL_PRINTOUT_STR, len, f);
 	}
 	else
 	{
 		len = ft_strlen(str);
-		if (f->precision >= 0 && len > f->precision)
+		if (f->precision >= 0 && len > (size_t)f->precision)
 			len = f->precision;
 		printed += print(str, len, f);
 	}
 	return (printed);
 }
 
-static int	print(const char *str, int len, t_format *f)
+static size_t	print(const char *str, size_t len, t_format *f)
 {
-	int	printed;
+	size_t	printed;
 
 	printed = 0;
 	if (!f->minus && f->width > len)
